Create interactive_demo.py
I came across this project and I thought it would be interesting to have an interactive demo to better visualize the effects of MLS across the different types of distortion.

I have created a more or less minimal UI that uses cv2 windows and no other packages to achieve this. The downside is that it's all done through hotkeys since the cv2 windows have no support for buttons.